Button Checkbox Label Textarea as styled and compos elements with stories
