--- src/components/Label/StyledLabel.js
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
-import colors from '../../assets/styles/colors';
-
-const StyledTextarea = styled.label`
-  width: 100%;
-  font-family: 'Lato';
-  font-weight: 600;
-  font-size: 13px;
-  color: ${colors.blueTxt};
-`;
-
-StyledTextarea.defaultProps = {
-  placeholder: null,
-};
-
-StyledTextarea.propTypes = {
-  placeholder: PropTypes.string,
-};
-
-export default StyledTextarea;
