--- src/components/Textarea/Textarea.js
@@ -1,29 +1,0 @@
-/**
- *
- * Textarea
- *
- */
-
-import React from 'react';
-import PropTypes from 'prop-types';
-import {
-  commonDefaultProps,
-  commonPropTypes,
-} from '../../commonPropTypes/input';
-import StyledTextArea from './StyledTextarea';
-
-function Textarea(props) {
-  return <StyledTextArea {...props} />;
-}
-
-Textarea.defaultProps = {
-  ...commonDefaultProps,
-  value: null,
-};
-
-Textarea.propTypes = {
-  ...commonPropTypes,
-  value: PropTypes.string,
-};
-
-export default Textarea;
