+++ src/styled/Label/index.js
@@ -1,0 +1,18 @@
+/**
+ *
+ * Label
+ *
+ */
+
+import styled from 'styled-components';
+import colors from '../../assets/styles/colors';
+
+const Label = styled.label`
+  width: 100%;
+  font-family: 'Lato';
+  font-weight: 600;
+  font-size: 13px;
+  color: ${colors.blueTxt};
+`;
+
+export default Label;
