+++ src/styled/Textarea/tests/Textarea.test.js
@@ -1,0 +1,11 @@
+// import React from 'react';
+// import { mount } from 'enzyme';
+// import { enzymeFind } from 'styled-components/test-utils';
+
+// import Textarea from '../index';
+
+describe('<Textarea />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(true);
+  });
+});
